Added cache for compiled selector functions
This is a very naïve implementation of a cache for compiled selector functions.
Profiling showed that indeed (and unsurprisingly) the 'eval' is very expensive.
I have to clue how to replace it by a JITable function, but this, together with
the fast path for _id-only queries cuts down the time spent in this function by
at least on order of magnitude for my app. Note that, lacking a benchmark suite
for meteor, these measurements are as unscientific as they get.

Quite possibly it is a good idea to add a counter to each function in the cache
and deleting little-used functions once in a while, but for now I have kept this
as simple as possible.